backend/requirements.txt
Django>=5.2.0
djangorestframework>=3.14.0
django-filter>=24.3
django-cors-headers>=4.3.0
Pillow>=11.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
requests>=2.31.0
yookassa>=3.0.0
python-telegram-bot>=21.0.0
supabase>=2.0.0
redis>=5.0.0
django-redis>=5.4.0
psutil>=7.0.0
openpyxl>=3.1.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
celery>=5.3.0
celery[redis]>=5.3.0
python-magic>=0.4.27

# FIX: Pin versions to avoid SSL_MODE_ACCEPT_MOVING_WRITE_BUFFER error
Twisted==24.10.0  # Downgrade from 25.5.0
pyOpenSSL==24.2.1  # Downgrade from 25.3.0
cryptography==43.0.3  # Compatible version

# Testing dependencies
pytest>=7.4.0
pytest-django>=4.7.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-xdist>=3.5.0
faker>=20.0.0
factory-boy>=3.3.0

# Scheduling dependencies
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
httpx>=0.25.0
python-dateutil>=2.8.2
pytz>=2023.3

# Notification channel dependencies
firebase-admin>=6.4.0
twilio>=8.10.0